requirements.txt
beautifulsoup4==4.8.2
chromedriver-binary==80.0.3987.16.0
dj-database-url==0.5.0
Django==3.0.2
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
gunicorn==20.0.4